v 1.5

+++ app/src/main/java/com/example/doannhom4_quanlythuvien/model/Book.java
@@ -8,16 +8,16 @@ public class Book implements Serializable {
     private String id, title, author, coverPhotoURL, link;
     private String type;
     private float rating;
-    private float timestamp;
+    private long timestamp;
 
     public Book() {
     }
 
-    public float getTimestamp() {
+    public long getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(float timestamp) {
+    public void setTimestamp(long timestamp) {
         this.timestamp = timestamp;
     }
